massive structural changes to allow any part of the toolchain to be switched in and out for exeuction purposes. (currently still keeps provaence data as a fixed request, but monday ill fix that)
