+++ spinn_front_end_common/interface/buffer_management/buffer_manager.py
@@ -47,6 +47,7 @@ import threading
 import logging
 import traceback
 import os
+import re
 
 
 logger = logging.getLogger(__name__)
@@ -64,7 +65,7 @@ class BufferManager(object):
     """
 
     def __init__(self, placements, tags, transceiver, report_states,
-                 application_folder_path, reload_interface):
+                 application_folder_path):
         """
 
         :param placements: The placements of the vertices
@@ -86,8 +87,8 @@ class BufferManager(object):
         # params used for reload purposes
         self._report_states = report_states
         self._application_folder_path = application_folder_path
-        self._reload_interface = reload_interface
         self._reload_buffer_file = dict()
+        self._reload_buffer_file_paths = dict()
 
         # Set of (ip_address, port) that are being listened to for the tags
         self._seen_tags = set()
@@ -157,13 +158,16 @@ class BufferManager(object):
 
         # if reload script is set up, sotre the buffers for future usage
         if self._report_states.transciever_report:
-            vertex_files = self._reload_interface.add_buffered_vertex(
-                vertex, tag,
-                self._placements.get_placement_of_subvertex(vertex))
-            for (region, filename) in vertex_files.iteritems():
+            for region in vertex.get_regions():
+                filename = "{}_{}".format(
+                    re.sub("[\"':]", "_", vertex.label), region)
                 file_path = os.path.join(
                     self._application_folder_path, filename)
-            self._reload_buffer_file[(vertex, region)] = open(file_path, "w")
+                self._reload_buffer_file[(vertex, region)] = \
+                    open(file_path, "w")
+                if vertex not in self._reload_buffer_file_paths:
+                    self._reload_buffer_file_paths[vertex] = dict()
+                self._reload_buffer_file_paths[vertex][region] = file_path
 
     def load_initial_buffers(self):
         """ Load the initial buffers for the senders using mem writes
@@ -410,3 +414,20 @@ class BufferManager(object):
         if self._report_states.transciever_report:
             for buffer_file in self._reload_buffer_file.itervalues():
                 buffer_file.close()
+
+    @property
+    def sender_vertices(self):
+        """
+        property method for getting the vertices which are buffered
+        :return:
+        """
+        return self._sender_vertices
+
+    @property
+    def reload_buffer_files(self):
+        """
+        property method for getting the file paths for each buffered region
+        for each sender vertex
+        :return:
+        """
+        return self._reload_buffer_file_paths
