+++ spinn_front_end_common/interface/interface_functions/front_end_common_chip_execute_data_specification.py
@@ -1,0 +1,5 @@
+
+class FrontEndCommonChipExecuteDataSpecification(object):
+
+    def __call__(self):
+        raise NotImplementedError
